+++ tests/e2e/__init__.py
@@ -1,0 +1,4 @@
+"""End-to-end tests for QuickScale."""
+from .support.command_runner import run_quickscale_command
+
+__all__ = ['run_quickscale_command']